Viewport: keep clip height in small z

+++ src/OpenLoco/include/OpenLoco/Viewport.hpp
@@ -250,7 +250,7 @@ namespace OpenLoco::Ui
         }
     }
 
-    int16_t getMaxClipHeight();
-    void setMaxClipHeight(int16_t maxClipHeight);
+    uint8_t getMaxClipHeight();
+    void setMaxClipHeight(uint8_t maxClipHeight);
     void resetMaxClipHeight();
 }
